improved getfame.pro

+++ getfame.pro
@@ -1,9 +1,10 @@
 
-----------------------------------------------------------------------
-----------------------------------------------------------------------
+
+----------------------------------------------------------------------------------------------------------------------
+---------------------------------------------------------------------------------------------------------------------
 --ERIK Summer 2024 / 2025 erik.soberg@ssb.no / eriksoeb@gmail.com       
 --Programmed in FAME 4GL
-------------------------------------------------------------------
+---------------------------------------------------------------------------------------------------------------------
 --jul2024 init version with json for jupiterlab R Python 
 --sep2024 takes fame argument, such as freq a, convert on; deci 1 semikolon separated in double quotes
 --oct 2024 can take comma separated list of wildcards
@@ -12,12 +13,12 @@
 --jun2025 Can use getfame -e with dbname'series
 --seriesnames prefixed with ' after dbchannel and seriesname
 --jun2025 Handeling identical seriesnames in different databases
+--and more:
+--jun2025 Handeling 2 similar databasenames if they are opened in different paths..by adding 2,3,4 to the openas names
+--jun2025 Adding frequiency for series and NC for formulas using the -n getfameseriesnames option
+--jun2025 Cleaning away old writerror procedure
 
-----------------------------------------------------------------------
-
-
-
-----------------------------------------------------------------------
+---------------------------------------------------------------------------------------------------------------------
 procedure $createwildlist
 argument fbase, wild, mpathfile
 
@@ -29,7 +30,7 @@ over on
  scalar mcount :numeric = 0
  series finnescase:string by case
  series argcase:string by case
- --scalar myliste :namelist = {} --phase out due to '
+ 
  series mycase:string by case --new
  scalar num_missing :numeric = 0
  scalar missings :string = "" --collects series not found
@@ -37,7 +38,7 @@ over on
 local scalar reststr :string = trim(wild)
 local scalar curwild :string = ""
 
---type wild
+
 local scalar <over on>  myledd:numeric = 1
 
 
@@ -64,14 +65,20 @@ end if
 
 --$debug "inside wildlist mbase as .."
 
-local scalar mbase :string ="MYFAMEBASE"
+local scalar mbase :string =""
 
 loop for i=1 to lastvalue(basecase)
---set mbase = "MYFAMEBASE"+string(i)
---set mbase = basecase[i]
-set mbase = $getdbas(basecase[i])
 
---tester hele listen for gyldig wildcard
+--$debug "mbase wild :" + basecase[i] +string(i)
+
+if i GT  1
+set mbase = $getdbas(basecase[i]) + string(i) --if databases with identical names different places I add a number to make unique
+else
+set mbase = $getdbas(basecase[i]) --the first databse keeps original name
+end if
+
+
+--tester hele listen for valid wildcards
 loop for w=1 to lastvalue( argcase)
 
 
@@ -84,13 +91,13 @@ loop for w=1 to lastvalue( argcase)
 --maa lope for alle basene i global liste 1 ... eller open.db
 
 loop for s in (wildlist(id(mbase),argcase[w]))
-	--write "wild TESTING FINNES: "+ name(s) to OUTFILE
+	
 	if exists(id(name(s)))
---orig ok		set finnescase[mcount] = trim(name(s))
+
 		set finnescase[mcount] = upper(mbase)+"'"+trim(name(s)) --obs new
 
 		set mcount = mcount +1
-		--write "JEG FINNES: "+ name(s) to OUTFILE
+		
 	else
 		set missings = missings +name(s)+" "
 		set num_missing = num_missing +1
@@ -110,7 +117,7 @@ end loop --base(r)
 
 OTHERWISE
 --$debug "l_err:" + lasterror
-	$writeerror_clean "Ups! check for valid wildcard-list: "+wild, fbase,mpathfile --+ lasterror
+	$writeerror_clean "Ups! check for valid wildcard-list: "+wild, fbase,mpathfile+ lasterror
 	$close_html_file
 	$exitfame
 END TRY
@@ -132,7 +139,7 @@ channel warning none
 close all
 
 
---$debug "getfamenames debug starts here"
+--$debug "getfamenames debug starts herei ------------------------------------------"+ now
 
 local scalar mpath:string ="$HOME/.GetFAME/"
 local scalar mfile:string ="getfamenames.json"
@@ -144,12 +151,15 @@ local scalar myclass:string ="series"
 local scalar mycreated:string ="created time"
 local scalar myupdated:string ="upd time"
 local scalar mydescr:string ="desc of object"
-local scalar myobs:string = "Hi" 
+local scalar myobs:string = "myobs" 
+local scalar myfreq:string = "myfreq" 
 
 
 $open_html_file mpathfile
 
 $openbase fbase
+
+--$debug "did open " + @open.db + "  s "+@search
 
 
 
@@ -163,7 +173,6 @@ $createwildlist fbase, wild, mpathfile
 
 
 try
---local scalar mylength:numeric = length  (myliste) 
 local scalar mylength:numeric = lastvalue (mycase) 
 local scalar mycount:numeric = 0  --new variable counting
 if mylength EQ 0.0 or missing(mylength)
@@ -192,29 +201,26 @@ image date value quarterly "<FYEAR>-<MZ>-01"
 
 replace NC "NC" 
 
---$debug "looping need more work when here"
---loop for x in (myliste) 
+
 loop for s=1 to lastvalue(mycase)  --ny 
 
 	set mycount = mycount+1
 	set sn = mycase[s] --seriesname
 
---$debug  "write seriesname: " + sn 
-
 
 execute "set myclass = class(id(name(" +sn+ ")))"
 
---$debug  "myclass: " + myclass 
-
 if myclass EQ "SERIES"
 execute "set myobs = observed(id(name(" +sn+ ")))"
+execute "set myfreq = freq(id(name(" +sn+ ")))"
 
-else if myclass EQ "FORMULA" 
-
+else if myclass EQ "FORMULA"
+set myfreq = "NC"  --will always return this anyway
 execute "set myobs = source(id(name(" +sn+ ")))"
 
-else
+else  --other objects will never occur due to wildcarding formulas and series only..
 set myobs="OBSproblemClass"
+set myfreq = "freqproblem" 
 
 end if
 
@@ -227,9 +233,16 @@ execute "set mydescr =  descr(" +sn +")"
 
 
 
+	write <crlf off> "{"+quote +"Name"+quote+":"+ quote + sn +quote +","        &&
+	    + quote+"Class"+quote+":"+quote+myclass+quote +","   &&
+	    + quote+"Observed"+quote+":"+quote+myobs+quote+","        &&
+	    + quote+"Freq"+quote+":"+quote+myfreq+quote+","        &&
+	    + quote+"Desc"+quote+":"+quote+replace(mydescr,quote,"'")+quote +"," &&
+	    + quote+"Created"+quote+":"+ quote+mycreated+quote+","+     &&
+	    + quote+"Updated"+quote+ ":"+ quote+myupdated +quote  +"}" to Outfile
 
 
-	write <crlf off> "{"+quote +"Name"+quote+":"+ quote + sn +quote +","            + quote+"Class"+quote+":"+quote+myclass+quote +","    +quote+"Observed"+quote+":"+quote+myobs+quote+","           + quote+"Desc"+quote+":"+quote+replace(mydescr,quote,"'")+quote +"," +quote+"Created"+quote+":"+ quote+mycreated+quote+","+     quote+"Updated"+quote+ ":"+ quote+myupdated +quote  +"}" to Outfile
+
 
 
 --oif (lastvalue(myserie) NE mydate) 
@@ -244,7 +257,7 @@ end loop
 $log "getfamenames", fbase, wild , ""
 
 otherwise 
-$writeerror "-- "+lasterror + " !!!", fbase
+$writeerror_clean "-- "+lasterror + " !", fbase,"NC"
 end try
 
 
@@ -267,7 +280,6 @@ store work; over on
 channel warning none
 ignore on
 deci auto
---$debug "defaults new run:::"
 end proc
 ------------------------------------------
 
@@ -314,12 +326,11 @@ image date value quarterly "<FYEAR>-<MZ>-01"
 
 try
 execute famearg
---|if exists(id(curve2)
+
 if exists(curve2)
 	execute "local new getfame_ser = "+curve2
 	--desc (getfame_ser) = id(name(curve2))    --just name is best, if lsum is used what is description
 	desc (getfame_ser) = curve2    --just name is best, if lsum is used what is description
-	--execute "desc (getfame_ser) =descr("+ curve2+"  )"
 
 else
 $writeerror_clean "Ups "+upper(curve2) +" Not Found", fbase, mpathfile
@@ -328,8 +339,6 @@ $exitfame
 end if
 
 
-
---	desc (mytest) = curve2   
 
 otherwise
 $writeerror_clean lasterror, fbase, mpathfile
@@ -342,7 +351,6 @@ end try
 if firstvalue(getfame_ser) NE NC 
 $getobs fbase, curve2, getfame_ser 
 else 
---$writeerror_clean "Series " + Curve2 + " No Values in Range", fbase, mpathfile
 $norange curve2,getfame_ser
 end if
 
@@ -418,7 +426,6 @@ try
 execute famearg
 
 otherwise
-$writeerror "  " + lasterror , fbase
 $writeerror_clean lasterror , fbase,mpathfile
 $close_html_file
 $exitfame
@@ -431,7 +438,6 @@ item type string off, namelist off, boolean off
 item type precision on, numeric on
 item alias on
 
-----
 --$debug "getmyfameseries "
 
 $createwildlist fbase, wild, mpathfile
@@ -450,7 +456,6 @@ if mylength EQ 0
 end if
 
 
---extra for founds
 
 
 $header fbase ,mpathfile
@@ -466,38 +471,29 @@ write quote+"Series"+quote+": [  " to OUTFILE
 --$debug "getmyfameseries looping "+string(mylength)
 
 set mysname =  "floopsname" 
---loop for x in (myliste) 
+
 loop for s =1 to lastvalue(mycase)  
 	set mycount = mycount +1
 
---set mysname =  string(name(x)) 
+
 set mysname =   mycase[mycount]
 
 --$debug "getmyfameseries looping "+mysname
 
-	--execute "local new mytest = "+name(x)
-	execute "local new myseries = "+mysname
-
-	--execute "desc (mytest) = desc(" +name(x) +")" 
+	
+	execute "local new myseries = "+mysname	
 	execute "desc (myseries) = desc(" +mysname +")" 
 
 --$debug "getmyfameseries looping executes.."
 
 desc(myseries) = replace( desc(myseries), quote, "'")
-        --write string(mylength) + " / " + string(mycount) to outfile --debug
+       
 
 if firstvalue(myseries) NE NC
---$debug "getmyfameseries getobs call:"+mysname + "    : " +name(mytest)
-	--$getobs fbase, name(x),  mytest
 	$getobs fbase, mysname,  myseries
 else
 
 $norange mysname, myseries
---write "{" +quote+"Name"+quote+ ": "+quote+string(upper(  mysname))+quote +"," to OUTFILE
---write quote+"Desc" +quote+": "+ quote+desc(mytest) + quote + ","  to OUTFILE
---write quote+"Daterange" +quote+": "+ quote+(replace(@date,"NULL","*")) + quote +","  to OUTFILE
---write quote+"Frequency" +quote+": "+ quote+freq(mytest) + quote + ","  to OUTFILE
---write quote+"Observations"+quote+":[ ]" +NEWLINE + "}" to Outfile
 
 end if
 	if mycount NE mylength
@@ -560,18 +556,18 @@ write quote+"Date"+quote+":"+quote+myday +"T"+now+quote to ERRORFILE
 write " } }"  to ERRORFILE
 
 close !ERRORFILE
-
-system("rm -f "+mpathfile)
-system("touch "+mpathfile)
-
+if mpathfile NE "NC"  --deleting the json file case its corrupted
+	system("rm -f "+mpathfile)
+	system("touch "+mpathfile)
+end if
 
 end proc
 -----------------------------------ny write error-------------------------
 
 
 
--------------------still in use ----------------------
-procedure $writeerror
+-------------------not in use after jun 4 2025----------------------
+procedure $writeerror_phasedot
 argument myerror, mybase
 
 --image date value annual "<FYEAR>-01-01"
@@ -584,7 +580,7 @@ local scalar myday:string = today ("<FYEAR>-<MZ>-<DZ>")
 width 30000
 open <kind text; acc overwrite>file ("$HOME/.GetFAME/fame.error") as ERRORFILE
 
-write "{" + quote+"Error"+quote+": {"  to ERRORFILE
+write "{" + quote+"error"+quote+": {"  to ERRORFILE
 write quote+"GetFameError"+quote+":"+quote+err2+quote + "," to ERRORFILE
 write quote+"Database"+quote+":"+quote+mybase+quote + "," to ERRORFILE
 write quote+"Date"+quote+":"+quote+myday +"T"+now+quote to ERRORFILE
@@ -600,18 +596,11 @@ end proc
 -------------------debug  use ----------------------
 procedure $debug
 argument myerror
-
- 
-
 local scalar myday:string = today ("<FYEAR>-<MZ>-<DZ>")
-
 width 30000
 open <kind text; acc append>file ("$HOME/.GetFAME/fame.debug") as ERRORFILE
-
 write myday+ " : " +myerror  to ERRORFILE
-
 close !ERRORFILE
-
 end proc
 -------------------------------------------------------------
 
@@ -626,8 +615,6 @@ over on
 local scalar restbase :string = trim(fbase)
 local scalar curbase :string = ""
 series basecase :string by case
---series openascase :string by case
---series <case 1 to *> openascase :string by case = "TESTAPI", "MYFAMEBASE1","DB2"
 local scalar myledd :numeric = 1
 
 
@@ -636,8 +623,6 @@ loop while not missing(location(restbase, ",")) and restbase NE ","
 set curbase =   trim(substring(restbase,1,location(restbase,",")-1))
 if  curbase NE ""  --duplicate commas
         set basecase[myledd] =  curbase
---       set openascase[myledd] = $getdbas(curbase) 
---
         set myledd = myledd+1
 end if
         set restbase=trim(substring(restbase, location(restbase,",")+1, length(restbase)))
@@ -646,20 +631,22 @@ end loop
 
 if (length(restbase) GT 0 ) and (trim(restbase) NE ",")
         set basecase[myledd] =  trim(restbase )
---        set openascase[myledd] = $getdbas(curbase) 
 end if
 
 
 loop for i = 1 to lastvalue(basecase)
 
-
+--$debug "openbase..  :" +basecase[i] + string(i)
 try
-execute "open <acc r> "+ quote+trim(basecase[i]) +quote + " as  "+$getdbas(basecase[i] )
+if i GT 1
+execute "open <acc r> "+ quote+trim(basecase[i]) +quote + " as  "+$getdbas(basecase[i]) +string(i) 
+else
+execute "open <acc r> "+ quote+trim(basecase[i]) +quote + " as  "+$getdbas(basecase[i]) 
+end if
 
---orig OK :execute "open <acc r> "+ quote+trim(basecase[i]) +quote + " as  myfamebase"+string(i)
 
 otherwise
-$writeerror "  " + lasterror, basecase[i]
+$writeerror_clean "Open:" + lasterror, basecase[i],"NC"
 $close_html_file
 $exitfame
 end try
@@ -688,13 +675,6 @@ set slashfree=dotfree
 if (location(dotfree,"/") NE NC )
 set slashfree =substring(dotfree, 1, location(dotfree,"/")-1)
 end if
---set bname = string(mbase)
-
-
---need to take away // paths $REFERTID/data/kpi_publ.db"
-
---$writeerror string(dotfree) ,  bname
---$writeerror mirror(string(slashfree)) ,  bname
 
 --withoit dot and slashes
 return mirror(trim(string(slashfree)))
@@ -719,22 +699,20 @@ scalar <over on>  stop_time  :precision =start_time --init to 0
 local scalar mynow :string = string(datefmt(created(start_time,secondly),"<YEAR>-<MZ>-<DZ>T<HHZ>:<MMZ>:<SSZ>"))
 
 
-
-
 try
 write "[{"+quote+"GetFAME_Json_Api" +quote+": "+ quote+"Erik.Soberg@ssb.no" + +quote+","  to OUTFILE
-write quote+"Version" +quote+": "+ quote+"Oslo-20250602" + quote+","  to OUTFILE
+write quote+"Version" +quote+": "+ quote+"Oslo-20250605" + quote+","  to OUTFILE
 
 write quote+"Executed" +quote+": "+ quote+ + mynow +quote+","  to OUTFILE
 write quote+"Famever"+quote+": "+quote+ string(@release)+quote+","  to OUTFILE
 write quote+"Database"+quote+": "+quote+ fbase+quote +"," to OUTFILE
 
-write quote+"Open"+quote+": "+quote+ @open.db+quote +"," to OUTFILE
+write quote+"Openas"+quote+": "+quote+ @open.db+quote +"," to OUTFILE
 
 --write quote+"Result"+quote+": "+quote+ replace(ofile,"iso","")+quote +"," to OUTFILE
 write quote+"Result"+quote+": "+quote+ ofile+quote +"," to OUTFILE
 otherwise
-$writeerror "  " + lasterror, fbase
+$writeerror_clean "  " + lasterror, fbase,"NC"
 end try
 end proc
 ---------------------------
@@ -755,10 +733,10 @@ replace ND "null"
 
 
 local scalar < over on> mydate:date
---local scalar < over on> myobs:numeric
+
 
 try
---write "{" +quote+"NameTEST"+quote+ ": "+quote+string(upper( myserie))+quote +"," to OUTFILE
+
 write "{" +quote+"Name"+quote+ ": "+quote+string(upper( curve))+quote +"," to OUTFILE
 write quote+"Desc" +quote+": "+ quote+desc(myserie) + quote + ","  to OUTFILE
 write quote+"Daterange" +quote+": "+ quote+(replace(@date,"NULL","*")) + quote +","  to OUTFILE
@@ -766,7 +744,7 @@ write quote+"Frequency" +quote+": "+ quote+freq(myserie) + quote + ","  to OUTFI
 
 
 otherwise
-$writeerror "  " + lasterror, fbase
+$writeerror_clean "  " + lasterror, fbase, "NC"
 end try
 
 write quote+"Observations"+quote+":[" to Outfile
@@ -786,7 +764,7 @@ end if
 
 
 otherwise
-$writeerror "  " + lasterror, fbase
+$writeerror_clean "  " + lasterror, fbase,"NC"
 end try
 end loop
 
@@ -821,7 +799,6 @@ end proc
 ---------------------------------------------------------------
 proc $open_html_file
 argument filename
---system("rm -f "+filename)
 open <kind text; acc overwrite>file (filename) as OUTFILE
 end proc
 ---------------------------------------------------------------
@@ -835,7 +812,7 @@ output terminal
 close !outfile
 close all
 otherwise
-$writeerror "  " + lasterror, "closing" --makes no sense with database 
+$writeerror_clean "  " + lasterror, "closing" ,"NC"--makes no sense with database 
 end try
 end proc
 ---------------------------------------------
@@ -851,10 +828,10 @@ end proc
 
 
 ------------------------------------------------------------------------------------------
---$epoc --Kudos to Jeff K for inspiring on this one
+--$epoc --Kudos to Jeff for help on this one
 --when having a timeseries system, you should never stop using the real date ..
 ------------------------------------------------------------------------------------------
-function $epoc --Kudos to Jeff K for inspiring on this one
+function $epoc --msec since 1970
 argument fdate
 return dateof(fdate,millisecondly,after,begin) - millisecondly(1)(1jan1970:0:0:0)
 end function
@@ -862,7 +839,7 @@ end function
 
 
 
-
+--Eof
